+++ jest.config.js
@@ -25,8 +25,8 @@ module.exports = {
   },
 
   // The glob patterns Jest uses to detect test files
-  testMatch: ["**/test/unit/**/?(*.)+(spec|test).js?(x)"],
-  //testMatch: ["**/test/unit/core/Plugins.spec.js"],
+  testMatch: ["<rootDir>/test/unit/**/*.spec.js"],
+  // testMatch: ["**/test/unit/**/Cli.spec.js"],
 
   // The test environment that will be used for testing
   testEnvironment: "node",