+++ test/unit/Core.mocks.js
@@ -36,15 +36,18 @@ class CoreMock {
         error: this._sandbox.stub(),
       },
       onChangeSettings: this._sandbox.stub().returns(() => {}),
+      onChangeAlerts: this._sandbox.stub().returns(() => {}),
       onChangeMocks: this._sandbox.stub().returns(() => {}),
       addRouter: this._sandbox.stub(),
       addSetting: this._sandbox.stub(),
       behaviors: {
-        ids: [],
-        currentFromCollection: "foo-current",
-        collection: "foo-behaviors-collection",
+        count: 0,
+        currentId: "foo-current",
       },
-      fixtures: {},
+      alerts: [],
+      fixtures: {
+        count: 0,
+      },
       serverError: null,
       _eventEmitter: {
         on: this._sandbox.stub(),