+++ test/unit/src/Cli.spec.js
@@ -10,6 +10,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 */
 
 const sinon = require("sinon");
+const chalk = require("chalk");
 
 const CoreMocks = require("../Core.mocks.js");
 const InquirerMocks = require("./Inquirer.mocks.js");
@@ -47,6 +48,10 @@ describe("Cli", () => {
       cli = new Cli(coreInstance);
       expect(coreInstance.addSetting.getCall(0).args[0].name).toEqual("cli");
     });
+
+    it("should have displayName", () => {
+      expect(cli.displayName).toEqual("@mocks-server/plugin-inquirer-cli");
+    });
   });
 
   describe("when initializated", () => {
@@ -61,6 +66,12 @@ describe("Cli", () => {
       await cli.init();
       expect(inquirerMocks.stubs.Inquirer.callCount).toEqual(0);
     });
+  });
+
+  describe("when settings are changed", () => {
+    beforeEach(async () => {
+      await cli.start();
+    });
 
     it("should start cli when core cli setting is true and cli was not started", async () => {
       expect.assertions(2);
@@ -69,12 +80,6 @@ describe("Cli", () => {
       });
       expect(inquirerMocks.stubs.inquirer.inquire.callCount).toEqual(1);
       expect(inquirerMocks.stubs.inquirer.inquire.getCall(0).args[0]).toEqual("main");
-    });
-  });
-
-  describe("when settings are changed", () => {
-    beforeEach(async () => {
-      await cli.start();
     });
 
     it("should refresh main menu when behavior option is changed and current screen is main menu", async () => {
@@ -204,6 +209,19 @@ describe("Cli", () => {
     });
   });
 
+  describe("when alerts are changed", () => {
+    beforeEach(async () => {
+      await cli.start();
+    });
+
+    it("should refresh main menu", async () => {
+      expect.assertions(2);
+      coreInstance.onChangeAlerts.getCall(0).args[0]();
+      expect(inquirerMocks.stubs.inquirer.inquire.callCount).toEqual(2);
+      expect(inquirerMocks.stubs.inquirer.inquire.getCall(1).args[0]).toEqual("main");
+    });
+  });
+
   describe("when started", () => {
     beforeEach(async () => {
       await cli.start();
@@ -240,7 +258,17 @@ describe("Cli", () => {
   });
 
   describe("when stopped", () => {
+    let removeChangeMocksSpy;
+    let removeChangeAlertsSpy;
+    let removeChangeSettingsSpy;
+
     beforeEach(async () => {
+      removeChangeMocksSpy = sinon.spy();
+      removeChangeAlertsSpy = sinon.spy();
+      removeChangeSettingsSpy = sinon.spy();
+      coreInstance.onChangeSettings.returns(removeChangeSettingsSpy);
+      coreInstance.onChangeAlerts.returns(removeChangeAlertsSpy);
+      coreInstance.onChangeMocks.returns(removeChangeMocksSpy);
       await cli.start();
     });
 
@@ -252,6 +280,15 @@ describe("Cli", () => {
       expect(inquirerMocks.stubs.inquirer.clearScreen.getCall(0).args[0]).toEqual({
         header: false,
       });
+    });
+
+    it("should remove onChange listeners", async () => {
+      expect.assertions(3);
+      inquirerMocks.reset();
+      await cli.stop();
+      expect(removeChangeSettingsSpy.callCount).toEqual(1);
+      expect(removeChangeMocksSpy.callCount).toEqual(1);
+      expect(removeChangeAlertsSpy.callCount).toEqual(1);
     });
 
     it("should not stop if it was already stopped", async () => {
@@ -437,32 +474,123 @@ describe("Cli", () => {
   });
 
   describe("when printing header", () => {
-    it("should print it as first element if server has an error", async () => {
-      const fooServerErrorMessage = "foo server error";
-      const fooServerError = new Error(fooServerErrorMessage);
-      coreInstance.serverError = fooServerError;
-      await cli.start();
-      expect(cli._header()[0]).toEqual(expect.stringContaining(fooServerErrorMessage));
-    });
-
-    it("should print server url as first element if server has not an error", async () => {
-      coreInstance.serverError = null;
+    it("should print server url as first element", async () => {
       await cli.start();
       expect(cli._header()[0]).toEqual(expect.stringContaining("Mocks server listening"));
     });
 
     it("should print localhost as host when it is 0.0.0.0", async () => {
-      coreInstance.serverError = null;
       coreInstance.settings.get.withArgs("host").returns("0.0.0.0");
       await cli.start();
       expect(cli._header()[0]).toEqual(expect.stringContaining("http://localhost"));
     });
 
     it("should print custom host as host", async () => {
-      coreInstance.serverError = null;
       coreInstance.settings.get.withArgs("host").returns("foo-host");
       await cli.start();
       expect(cli._header()[0]).toEqual(expect.stringContaining("http://foo-host"));
+    });
+
+    it("should print delay in yellow if is greater than 0", async () => {
+      coreInstance.settings.get.withArgs("delay").returns(1000);
+      await cli.start();
+      expect(cli._header()[1]).toEqual(expect.stringContaining(chalk.yellow("1000")));
+    });
+
+    it("should print delay in green if is equal to 0", async () => {
+      coreInstance.settings.get.withArgs("delay").returns(0);
+      await cli.start();
+      expect(cli._header()[1]).toEqual(expect.stringContaining(chalk.green("0")));
+    });
+
+    it("should print behaviors in red if are equal to 0", async () => {
+      coreInstance.behaviors.count = 0;
+      await cli.start();
+      expect(cli._header()[2]).toEqual(expect.stringContaining(chalk.red("0")));
+    });
+
+    it("should print behaviors in green if are greater than 0", async () => {
+      coreInstance.behaviors.count = 10;
+      await cli.start();
+      expect(cli._header()[2]).toEqual(expect.stringContaining(chalk.green("10")));
+    });
+
+    it("should print current behavior in red if it is null", async () => {
+      coreInstance.behaviors.currentId = null;
+      await cli.start();
+      expect(cli._header()[3]).toEqual(expect.stringContaining(chalk.red("-")));
+    });
+
+    it("should print current behavior in green if it is defined", async () => {
+      coreInstance.behaviors.currentId = "foo";
+      await cli.start();
+      expect(cli._header()[3]).toEqual(expect.stringContaining(chalk.green("foo")));
+    });
+
+    it("should print current fixtures in red if there are less than 1", async () => {
+      coreInstance.fixtures.count = 0;
+      await cli.start();
+      expect(cli._header()[4]).toEqual(expect.stringContaining(chalk.red("0")));
+    });
+
+    it("should print current fixtures in green if there are less than 1", async () => {
+      coreInstance.fixtures.count = 10;
+      await cli.start();
+      expect(cli._header()[4]).toEqual(expect.stringContaining(chalk.green("10")));
+    });
+
+    it("should print watch in yellow if it is disabled", async () => {
+      coreInstance.settings.get.withArgs("watch").returns(false);
+      await cli.start();
+      expect(cli._header()[6]).toEqual(expect.stringContaining(chalk.yellow("false")));
+    });
+
+    it("should print watch in yellow if it is enabled", async () => {
+      coreInstance.settings.get.withArgs("watch").returns(true);
+      await cli.start();
+      expect(cli._header()[6]).toEqual(expect.stringContaining(chalk.green("true")));
+    });
+  });
+
+  describe("when printing alerts", () => {
+    it("should not display alerts if core alerts are empty", async () => {
+      coreInstance.alerts = [];
+      await cli.start();
+      expect(cli._alertsHeader().length).toEqual(0);
+    });
+
+    it("should display provided alert in yellow when it has no error", async () => {
+      expect.assertions(2);
+      coreInstance.alerts = [
+        {
+          message: "foo message",
+        },
+      ];
+      await cli.start();
+      expect(cli._alertsHeader()[0]).toEqual(expect.stringContaining("Warning"));
+      expect(cli._alertsHeader()[0]).toEqual(expect.stringContaining(chalk.yellow("foo message")));
+    });
+
+    it("should display provided alert in red when it has error", async () => {
+      expect.assertions(2);
+      coreInstance.alerts = [
+        {
+          message: "foo message",
+          error: {
+            message: "Foo error message",
+            stack: "Testing stack\nTesting stack 2\nTesting stack 3\nTesting stack 4",
+          },
+        },
+      ];
+      await cli.start();
+      expect(cli._alertsHeader()[0]).toEqual(expect.stringContaining("Error"));
+      expect(cli._alertsHeader()[0]).toEqual(
+        expect.stringContaining(
+          chalk.red(
+            `foo message: Foo error message\n         Testing stack\n         Testing stack 2\n         Testing stack 3...`
+          )
+        )
+      );
     });
   });
 
@@ -478,23 +606,6 @@ describe("Cli", () => {
 
     it("should exit logs mode", async () => {
       expect(inquirerMocks.stubs.inquirer.exitLogsMode.callCount).toEqual(2);
-    });
-  });
-
-  describe("stopListeningServerWatch method", () => {
-    it("should remove load:mocks listener if cli has been started", async () => {
-      const spy = sandbox.spy();
-      coreInstance.onChangeMocks.returns(spy);
-      await cli.start();
-      cli.stopListeningServerWatch();
-      expect(spy.callCount).toEqual(1);
-    });
-
-    it("should not remove load:mocks listener if cli has not been started", () => {
-      const spy = sandbox.spy();
-      coreInstance.onChangeMocks.returns(spy);
-      cli.stopListeningServerWatch();
-      expect(spy.callCount).toEqual(0);
     });
   });
 });